app/requirements.txt
flask==2.3.3
requests==2.31.0
mysql-connector-python==8.1.0
gunicorn==21.2.0
prometheus-client==0.17.1